+++ browser.js
@@ -16,6 +16,7 @@
 (function () {
 'use strict';
 
+window.__wctUseNpm = false;
 /**
  * @param {function()} callback A function to call when the active web component
  *     frameworks have loaded.
@@ -455,9 +456,6 @@ ChildRunner.prototype.signalRunComplete = function signalRunComplete(error) {
   this.onRunComplete = null;
 };
 
-var useNpm = document.currentScript.src.match(/wct-browser-legacy\/browser.js/) ||
-  (new URL(document.currentScript.src)).search.match(/[?&]npm=true/);
-
 /**
  * The global configuration state for WCT's browser client.
  */
@@ -467,7 +465,7 @@ var _config = {
    *
    * Paths are relative to `scriptPrefix`.
    */
-  environmentScripts: useNpm ?
+  environmentScripts: !!window.__wctUseNpm ?
     [
       'stacky/browser.js',
       'async/lib/async.js',
@@ -476,7 +474,8 @@ var _config = {
       'chai/chai.js',
       '@polymer/sinonjs/sinon.js',
       'sinon-chai/lib/sinon-chai.js',
-      'accessibility-developer-tools/dist/js/axs_testing.js'
+      'accessibility-developer-tools/dist/js/axs_testing.js',
+      '@polymer/test-fixture/test-fixture.js'
     ] : [
       'stacky/browser.js',
       'async/lib/async.js',
@@ -488,12 +487,7 @@ var _config = {
       'accessibility-developer-tools/dist/js/axs_testing.js'
     ],
 
-  environmentImports: useNpm ?
-    [
-      '@polymer/test-fixture/test-fixture.html'
-    ] : [
-      'test-fixture/test-fixture.html'
-    ],
+  environmentImports: !!window.__wctUseNpm ? [] : ['test-fixture/test-fixture.html'],
 
   /** Absolute root for client scripts. Detected in `setup()` if not set. */
   root: null,
@@ -1303,9 +1297,8 @@ function _injectPrototype(klass, prototype) {
  */
 function loadSync() {
   debug('Loading environment scripts:');
-  var a11ySuite =
-    document.currentScript.src.match(/wct-browser-legacy\/browser.js/) ?
-      'wct-browser-legacy/a11ySuite.js' : 'web-component-tester/data/a11ySuite.js';
+  var a11ySuite = !!window.__wctUseNpm ?
+    'wct-browser-legacy/a11ySuite.js' : 'web-component-tester/data/a11ySuite.js';
   var scripts = get('environmentScripts');
   var a11ySuiteWillBeLoaded = window.__generatedByWct || scripts.indexOf(a11ySuite) > -1;
   if (!a11ySuiteWillBeLoaded) {