+++ runner/config.ts
@@ -30,7 +30,7 @@ const HOME_DIR = path.resolve(
 const JSON_MATCHER = 'wct.conf.json';
 const CONFIG_MATCHER = 'wct.conf.*';
 
-export type Browser = string | {browserName: string, platform: string};
+export type Browser = string|{browserName: string, platform: string};
 
 export interface Config {
   suites?: string[];
@@ -61,6 +61,7 @@ export interface Config {
     _servers?: {variant: string, url: string}[];
   };
   npm?: boolean;
+  packageName?: string;
   skipPlugins?: string[];
   sauce?: {};
   remote?: {};
@@ -72,6 +73,50 @@ export interface Config {
   configFile?: string;
   /** A deprecated option */
   browsers?: Browser[]|Browser;
+}
+
+/**
+ * config helper: A basic function to synchronously read JSON,
+ * log any errors, and return null if no file or invalid JSON
+ * was found.
+ */
+function readJsonSync(filename: string, dir?: string): any|null {
+  const configPath = path.resolve(dir || '', filename);
+  let config: any;
+  try {
+    config = fs.readFileSync(configPath, 'utf-8');
+  } catch (e) {
+    return null;
+  }
+  try {
+    return JSON.parse(config);
+  } catch (e) {
+    console.error(`Could not parse ${configPath} as JSON`);
+    console.error(e);
+  }
+  return null;
+}
+
+
+/**
+ * Determines the package name by reading from the following sources:
+ *
+ * 1. `options.packageName`
+ * 2. bower.json or package.json, depending on options.npm
+ */
+export function getPackageName(options: Config): string|undefined {
+  if (options.packageName) {
+    return options.packageName;
+  }
+  const manifestName = (options.npm ? 'package.json' : 'bower.json');
+  const manifest = readJsonSync(manifestName, options.root);
+  if (manifest !== null) {
+    return manifest.name;
+  }
+  const basename = path.basename(options.root);
+  console.warn(
+      `no ${manifestName} found, defaulting to packageName=${basename}`);
+  return basename;
 }
 
 // The full set of options, as a reference.
@@ -164,7 +209,8 @@ export function defaults(): Config {
   };
 }
 
-/** nomnom configuration for command line arguments.
+/**
+ * nomnom configuration for command line arguments.
  *
  * This might feel like duplication with `defaults()`, and out of place (why not
  * in `cli.js`?). But, not every option matches a configurable value, and it is
@@ -473,9 +519,9 @@ function expandDeprecated(context: Context) {
         'The --browsers flag/option is deprecated. Please use ' +
             '--local and --sauce instead, or configure via plugins.' +
             '[local|sauce].browsers.');
-    const fragment: {plugins: {[name: string]: {browsers?: Browser[]}}} = {
-      plugins: {sauce: {}, local: {}}
-    };
+    const fragment: {
+      plugins: {[name: string]: {browsers?: Browser[]}}
+    } = {plugins: {sauce: {}, local: {}}};
     fragments.push(fragment);
 
     for (const browser of browsers) {